requirements.txt
pytelegrambotapi==4.14.0
pymongo==4.6.1
python-dotenv==1.0.0
